fixed budget storing

+++ app/src/main/java/com/papramaki/papramaki/models/Expenditure.java
@@ -4,22 +4,22 @@ import java.util.Date;
 
 public class Expenditure {
 
-    private float amount;
+    private double amount;
     private String category;
     private Date date;
 
-    public Expenditure(float amount, String category, Date date) {
+    public Expenditure(double amount, String category, Date date) {
         this.amount = amount;
         this.category = category;
         this.date = date;
     }
 
 
-    public float getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
